refactor project dir
