+++ src/bare/arch/aarch64/page_table.rs
@@ -1,6 +1,6 @@
 use aarch64_cpu::registers::{Writeable, TTBR0_EL1};
 
-use crate::{PhysAddr, PhysPage, VirtAddr, VirtPage};
+use crate::addr::{PhysAddr, PhysPage, VirtAddr, VirtPage};
 use crate::{MappingFlags, PageTable, PTE, TLB};
 use crate::bit;
