+++ src/bare/arch/aarch64/pl011.rs
@@ -3,7 +3,8 @@
 use arm_pl011::pl011::Pl011Uart;
 use irq_safety::MutexIrqSafe;
 
-use crate::{PhysAddr, DebugConsole};
+use crate::addr::PhysAddr;
+use crate::debug::DebugConsole;
 
 const UART_BASE: PhysAddr = PhysAddr(0x0900_0000);
