+++ src/bare/arch/loongarch64/sigtrx.rs
@@ -26,7 +26,7 @@ static mut TRX_STEP: [[PTE; PageTable::PTE_NUM_IN_PAGE]; 2] =
 pub fn init() {
     unsafe {
         TRX_STEP[0][0] = PTE::from_addr(
-            crate::PhysAddr(_sigreturn as usize & !VIRT_ADDR_START),
+            crate::addr::PhysAddr(_sigreturn as usize & !VIRT_ADDR_START),
             MappingFlags::URX.into(),
         );
         TRX_STEP[1][0] = PTE(TRX_STEP.as_ptr() as usize & !VIRT_ADDR_START);
