+++ src/bare/arch/riscv64/page_table/mod.rs
@@ -5,7 +5,7 @@ use core::arch::riscv64::sfence_vma;
 
 pub use sv39::*;
 
-use crate::VirtAddr;
+use crate::addr::VirtAddr;
 use crate::TLB;
 
 /// TLB operations
